+++ spark3/src/main/java/org/apache/iceberg/spark/source/SparkBatchScan.java
@@ -20,31 +20,49 @@
 package org.apache.iceberg.spark.source;
 
 import java.io.IOException;
+import java.io.Serializable;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.apache.iceberg.CombinedScanTask;
+import org.apache.iceberg.FileFormat;
 import org.apache.iceberg.FileScanTask;
 import org.apache.iceberg.Schema;
+import org.apache.iceberg.SchemaParser;
 import org.apache.iceberg.SnapshotSummary;
 import org.apache.iceberg.Table;
+import org.apache.iceberg.TableProperties;
 import org.apache.iceberg.TableScan;
 import org.apache.iceberg.encryption.EncryptionManager;
 import org.apache.iceberg.exceptions.RuntimeIOException;
 import org.apache.iceberg.expressions.Expression;
+import org.apache.iceberg.hadoop.HadoopInputFile;
+import org.apache.iceberg.hadoop.Util;
 import org.apache.iceberg.io.CloseableIterable;
 import org.apache.iceberg.io.FileIO;
+import org.apache.iceberg.mapping.NameMapping;
+import org.apache.iceberg.mapping.NameMappingParser;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
 import org.apache.iceberg.spark.Spark3Util;
 import org.apache.iceberg.spark.SparkSchemaUtil;
 import org.apache.iceberg.util.PropertyUtil;
 import org.apache.iceberg.util.TableScanUtil;
 import org.apache.spark.broadcast.Broadcast;
+import org.apache.spark.sql.catalyst.InternalRow;
+import org.apache.spark.sql.connector.read.Batch;
+import org.apache.spark.sql.connector.read.InputPartition;
+import org.apache.spark.sql.connector.read.PartitionReader;
+import org.apache.spark.sql.connector.read.PartitionReaderFactory;
+import org.apache.spark.sql.connector.read.Scan;
 import org.apache.spark.sql.connector.read.Statistics;
 import org.apache.spark.sql.connector.read.SupportsFileFilter;
+import org.apache.spark.sql.connector.read.SupportsReportStatistics;
+import org.apache.spark.sql.types.StructType;
 import org.apache.spark.sql.util.CaseInsensitiveStringMap;
+import org.apache.spark.sql.vectorized.ColumnarBatch;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -55,24 +73,40 @@ import static org.apache.iceberg.TableProperties.SPLIT_OPEN_FILE_COST_DEFAULT;
 import static org.apache.iceberg.TableProperties.SPLIT_SIZE;
 import static org.apache.iceberg.TableProperties.SPLIT_SIZE_DEFAULT;
 
-class SparkBatchScan extends BaseBatchScan implements SupportsFileFilter {
+class SparkBatchScan implements Scan, Batch, SupportsFileFilter, SupportsReportStatistics {
   private static final Logger LOG = LoggerFactory.getLogger(SparkBatchScan.class);
 
+  private final Table table;
+  private final boolean caseSensitive;
+  private final boolean localityPreferred;
+  private final Schema expectedSchema;
   private final List<Expression> filterExpressions;
   private final boolean ignoreResiduals;
   private final Long snapshotId;
   private final Long startSnapshotId;
   private final Long endSnapshotId;
   private final Long asOfTimestamp;
+  private final Long splitSize;
+  private final Integer splitLookback;
+  private final Long splitOpenFileCost;
+  private final Broadcast<FileIO> io;
+  private final Broadcast<EncryptionManager> encryptionManager;
+  private final boolean batchReadsEnabled;
+  private final int batchSize;
 
   // lazy cache of tasks
+  private StructType readSchema = null;
   private List<FileScanTask> files = null;
   private List<CombinedScanTask> tasks = null;
 
   SparkBatchScan(Table table, Broadcast<FileIO> io, Broadcast<EncryptionManager> encryption, boolean caseSensitive,
                  Schema expectedSchema, List<Expression> filters, boolean ignoreResiduals,
                  CaseInsensitiveStringMap options) {
-    super(table, expectedSchema, io, encryption, caseSensitive, options);
+    this.table = table;
+    this.io = io;
+    this.encryptionManager = encryption;
+    this.caseSensitive = caseSensitive;
+    this.expectedSchema = expectedSchema;
     this.filterExpressions = filters;
     this.ignoreResiduals = ignoreResiduals;
     this.snapshotId = Spark3Util.propertyAsLong(options, "snapshot-id", null);
@@ -94,6 +128,14 @@ class SparkBatchScan extends BaseBatchScan implements SupportsFileFilter {
     } else if (startSnapshotId == null && endSnapshotId != null) {
       throw new IllegalArgumentException("Cannot only specify option end-snapshot-id to do incremental scan");
     }
+
+    // look for split behavior overrides in options
+    this.splitSize = Spark3Util.propertyAsLong(options, "split-size", null);
+    this.splitLookback = Spark3Util.propertyAsInt(options, "lookback", null);
+    this.splitOpenFileCost = Spark3Util.propertyAsLong(options, "file-open-cost", null);
+    this.localityPreferred = Spark3Util.isLocalityEnabled(io.value(), table.location(), options);
+    this.batchReadsEnabled = Spark3Util.isVectorizationEnabled(table.properties(), options);
+    this.batchSize = Spark3Util.batchSize(table.properties(), options);
   }
 
   public Long snapshotId() {
@@ -111,6 +153,64 @@ class SparkBatchScan extends BaseBatchScan implements SupportsFileFilter {
     files = files().stream()
         .filter(file -> locations.contains(file.file().path().toString()))
         .collect(Collectors.toList());
+  }
+
+  @Override
+  public Batch toBatch() {
+    return this;
+  }
+
+  @Override
+  public StructType readSchema() {
+    if (readSchema == null) {
+      this.readSchema = SparkSchemaUtil.convert(expectedSchema);
+    }
+    return readSchema;
+  }
+
+  @Override
+  public InputPartition[] planInputPartitions() {
+    String tableSchemaString = SchemaParser.toJson(table.schema());
+    String expectedSchemaString = SchemaParser.toJson(expectedSchema);
+    String nameMappingString = table.properties().get(TableProperties.DEFAULT_NAME_MAPPING);
+
+    List<CombinedScanTask> scanTasks = tasks();
+    InputPartition[] readTasks = new InputPartition[scanTasks.size()];
+    for (int i = 0; i < scanTasks.size(); i++) {
+      readTasks[i] = new ReadTask(
+          scanTasks.get(i), tableSchemaString, expectedSchemaString, nameMappingString, io, encryptionManager,
+          caseSensitive, localityPreferred);
+    }
+
+    return readTasks;
+  }
+
+  @Override
+  public PartitionReaderFactory createReaderFactory() {
+    boolean allParquetFileScanTasks =
+        tasks().stream()
+            .allMatch(combinedScanTask -> !combinedScanTask.isDataTask() && combinedScanTask.files()
+                .stream()
+                .allMatch(fileScanTask -> fileScanTask.file().format().equals(
+                    FileFormat.PARQUET)));
+
+    boolean allOrcFileScanTasks =
+        tasks().stream()
+            .allMatch(combinedScanTask -> !combinedScanTask.isDataTask() && combinedScanTask.files()
+                .stream()
+                .allMatch(fileScanTask -> fileScanTask.file().format().equals(
+                    FileFormat.ORC)));
+
+    boolean atLeastOneColumn = expectedSchema.columns().size() > 0;
+
+    boolean onlyPrimitives = expectedSchema.columns().stream().allMatch(c -> c.type().isPrimitiveType());
+
+    boolean hasNoDeleteFiles = tasks().stream().noneMatch(TableScanUtil::hasDeletes);
+
+    boolean readUsingBatch = batchReadsEnabled && hasNoDeleteFiles && (allOrcFileScanTasks ||
+        (allParquetFileScanTasks && atLeastOneColumn && onlyPrimitives));
+
+    return new ReaderFactory(readUsingBatch ? batchSize : 0);
   }
 
   @Override
@@ -174,8 +274,7 @@ class SparkBatchScan extends BaseBatchScan implements SupportsFileFilter {
         snapshotId, startSnapshotId, endSnapshotId, asOfTimestamp);
   }
 
-  @Override
-  protected List<CombinedScanTask> tasks() {
+  private List<CombinedScanTask> tasks() {
     if (tasks == null) {
       Map<String, String> props = table.properties();
 
@@ -266,5 +365,124 @@ class SparkBatchScan extends BaseBatchScan implements SupportsFileFilter {
     return String.format(
         "IcebergScan(table=%s, type=%s, filters=%s, caseSensitive=%s)",
         table, expectedSchema.asStruct(), filterExpressions, caseSensitive);
+  }
+
+  private static class ReaderFactory implements PartitionReaderFactory {
+    private final int batchSize;
+
+    private ReaderFactory(int batchSize) {
+      this.batchSize = batchSize;
+    }
+
+    @Override
+    public PartitionReader<InternalRow> createReader(InputPartition partition) {
+      if (partition instanceof ReadTask) {
+        return new RowReader((ReadTask) partition);
+      } else {
+        throw new UnsupportedOperationException("Incorrect input partition type: " + partition);
+      }
+    }
+
+    @Override
+    public PartitionReader<ColumnarBatch> createColumnarReader(InputPartition partition) {
+      if (partition instanceof ReadTask) {
+        return new BatchReader((ReadTask) partition, batchSize);
+      } else {
+        throw new UnsupportedOperationException("Incorrect input partition type: " + partition);
+      }
+    }
+
+    @Override
+    public boolean supportColumnarReads(InputPartition partition) {
+      return batchSize > 1;
+    }
+  }
+
+  private static class RowReader extends RowDataReader implements PartitionReader<InternalRow> {
+    RowReader(ReadTask task) {
+      super(task.task, task.tableSchema(), task.expectedSchema(), task.nameMappingString, task.io(), task.encryption(),
+          task.isCaseSensitive());
+    }
+  }
+
+  private static class BatchReader extends BatchDataReader implements PartitionReader<ColumnarBatch> {
+    BatchReader(ReadTask task, int batchSize) {
+      super(task.task, task.expectedSchema(), task.nameMappingString, task.io(), task.encryption(),
+          task.isCaseSensitive(), batchSize);
+    }
+  }
+
+  private static class ReadTask implements InputPartition, Serializable {
+    private final CombinedScanTask task;
+    private final String tableSchemaString;
+    private final String expectedSchemaString;
+    private final String nameMappingString;
+    private final Broadcast<FileIO> io;
+    private final Broadcast<EncryptionManager> encryptionManager;
+    private final boolean caseSensitive;
+
+    private transient Schema tableSchema = null;
+    private transient Schema expectedSchema = null;
+    private transient NameMapping nameMapping = null;
+    private transient String[] preferredLocations = null;
+
+    ReadTask(CombinedScanTask task, String tableSchemaString, String expectedSchemaString, String nameMappingString,
+        Broadcast<FileIO> io, Broadcast<EncryptionManager> encryptionManager, boolean caseSensitive,
+        boolean localityPreferred) {
+      this.task = task;
+      this.tableSchemaString = tableSchemaString;
+      this.expectedSchemaString = expectedSchemaString;
+      this.nameMappingString = nameMappingString;
+      this.io = io;
+      this.encryptionManager = encryptionManager;
+      this.caseSensitive = caseSensitive;
+      if (localityPreferred) {
+        this.preferredLocations = Util.blockLocations(io.value(), task);
+      } else {
+        this.preferredLocations = HadoopInputFile.NO_LOCATION_PREFERENCE;
+      }
+    }
+
+    @Override
+    public String[] preferredLocations() {
+      return preferredLocations;
+    }
+
+    public Collection<FileScanTask> files() {
+      return task.files();
+    }
+
+    public FileIO io() {
+      return io.value();
+    }
+
+    public EncryptionManager encryption() {
+      return encryptionManager.value();
+    }
+
+    public boolean isCaseSensitive() {
+      return caseSensitive;
+    }
+
+    private Schema tableSchema() {
+      if (tableSchema == null) {
+        this.tableSchema = SchemaParser.fromJson(tableSchemaString);
+      }
+      return tableSchema;
+    }
+
+    private Schema expectedSchema() {
+      if (expectedSchema == null) {
+        this.expectedSchema = SchemaParser.fromJson(expectedSchemaString);
+      }
+      return expectedSchema;
+    }
+
+    private NameMapping nameMapping() {
+      if (nameMapping == null) {
+        this.nameMapping = NameMappingParser.fromJson(nameMappingString);
+      }
+      return nameMapping;
+    }
   }
 }
